feat(shaderlab): fix typo and add comments
