+++ packages/design/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./physics/index";
 export * from "./renderingHardwareInterface/index";
-export * from "./shaderLab";
+export * from "./shader-lab";
 export type { IClone } from "./IClone";
