+++ packages/shader-lab/scripts/gen_diagram.ts
