+++ packages/shader-lab/src/index.ts
@@ -1,0 +1,6 @@
+import { ShaderParser } from "./parser";
+import ShaderVisitor, { parser } from "./visitor";
+
+export * from "./ShaderLab";
+export { parseShader } from "./utils";
+export { ShaderParser, parser, ShaderVisitor };
