+++ packages/shader-lab/src/parser/rules/assignment.ts
