+++ packages/shader-lab/src/parser/rules/common.ts
