+++ packages/shader-lab/src/parser/rules/function/macro.ts
