+++ packages/shader-lab/src/parser/rules/function/statement.ts
