+++ packages/shader-lab/src/parser/rules/shader.ts
