+++ packages/shader-lab/src/parser/rules/struct.ts
